+++ bank-feeds/RELEASES.md
@@ -667,3 +667,13 @@ Based on:
 - [typescript v2.2.0] bank-feeds
 ### Releases
 - [NPM v2.2.0] https://www.npmjs.com/package/@codat/bank-feeds/v/2.2.0 - bank-feeds
+
+## 2023-11-06 01:29:42
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Bank-Feeds.yaml
+- Speakeasy CLI 1.112.1 (2.173.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v2.3.0] bank-feeds
+### Releases
+- [NPM v2.3.0] https://www.npmjs.com/package/@codat/bank-feeds/v/2.3.0 - bank-feeds